Deleted the createjob.js and editjobs.js and paste it their code to their page.js respective folders
